+++ CHANGELOG.md
@@ -78,6 +78,7 @@
 * Create Jest/RTL test for `RenderPermissions`. Refs UIU-2267.
 * Create Jest/RTL test for `PermissionSetDetails`. Refs UIU-2394.
 * Create Jest/RTL test for `PermissionsList`. Refs UIU-2374.
+* Create Jest/RTL test for `ProxyPermissions`. Refs UIU-2275.
 
 ## [6.1.0](https://github.com/folio-org/ui-users/tree/v6.1.0) (2021-06-18)
 [Full Changelog](https://github.com/folio-org/ui-users/compare/v6.0.0...v6.1.0)